autentikasi untuk aplikasi selesai

+++ Project/bersama/PemrogramanWeb/AryaYasirReza/KasirTopUp/view/admin/member.php
@@ -33,8 +33,14 @@
                             <td class="text-center">
                                 <a href="update_member.php?id=<?= $data['memberId'] ?>"
                                    class="btn btn-primary">Edit</a>
-                                <a href="../../routes/delete_member.php?id=<?= $data['memberId']; ?>"
-                                   class="btn btn-danger">Hapus</a>
+                                <?php if (
+                                    ($data['memberId'] != '6673d9722e552') &&
+                                    ($data['memberId'] != '6673d94bf37e4') &&
+                                    ($data['memberId'] != '6671dcad0f079')
+                                ) { ?>
+                                    <a href="../../routes/delete_member.php?id=<?= $data['memberId']; ?>"
+                                       class="btn btn-danger">Hapus</a>
+                                <?php } ?>
                             </td>
                         </tr>
                     <?php } ?>
